@@ -13,7 +13,7 @@ categories = [
 keywords = ["opentelemetry", "jaeger", "tracing", "async"]
 license = "Apache-2.0"
 edition = "2021"
-rust-version = "1.56"
+rust-version = "1.57"
 
 [package.metadata.docs.rs]
 all-features = true
