@@ -76,13 +76,10 @@ struct AppState {
 
 #[tokio::main]
 pub async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
-    let controller = controllers::basic(
-        processors::factory(
-            selectors::simple::histogram([1.0, 2.0, 5.0, 10.0, 20.0, 50.0]),
-            aggregation::cumulative_temporality_selector(),
-        )
-        .with_memory(true),
-    )
+    let controller = controllers::basic(processors::factory(
+        selectors::simple::histogram([1.0, 2.0, 5.0, 10.0, 20.0, 50.0]),
+        aggregation::cumulative_temporality_selector(),
+    ))
     .build();
 
     let exporter = opentelemetry_prometheus::exporter(controller).init();
@@ -124,7 +121,7 @@ pub async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
 
     let server = Server::bind(&addr).serve(make_svc);
 
-    println!("Listening on http://{}", addr);
+    println!("Listening on http://{addr}");
 
     server.await?;
 
