@@ -15,7 +15,7 @@
 //! let _span = tracer
 //!     .span_builder("span-name")
 //!     .with_attributes(vec![
-//!         semcov::trace::NET_PEER_IP.string("10.0.0.1"),
+//!         semcov::trace::NET_PEER_NAME.string("example.org"),
 //!         semcov::trace::NET_PEER_PORT.i64(80),
 //!     ])
 //!     .start(&tracer);
